+++ breeze-http/src/main/kotlin/com/windea/breezeframework/http/HttpConfig.kt
@@ -9,15 +9,16 @@ import javax.net.ssl.*
 /**
  * Http的配置。
  */
-data class HttpConfig(
-	val cookieHandler: CookieHandler? = null,
-	val connectTimeout: Duration? = null,
-	val sslContext: SSLContext? = null,
-	val sslParameters: SSLParameters? = null,
-	val executor: Executor? = null,
-	val followRedirects: HttpClient.Redirect? = null,
-	val version: HttpClient.Version? = null,
-	val priority: Int? = null,
-	val proxy: ProxySelector? = null,
-	val authenticator: Authenticator? = null
-)
+class HttpConfig {
+	var cookieHandler: CookieHandler? = null
+	var connectTimeout: Duration? = null
+	var sslContext: SSLContext? = null
+	var sslParameters: SSLParameters? = null
+	var executor: Executor? = null
+	var followRedirects: HttpClient.Redirect? = null
+	var version: HttpClient.Version? = null
+	var priority: Int? = null
+	var proxy: ProxySelector? = null
+	var authenticator: Authenticator? = null
+	var encoding: String = "UTF-8"
+}